Move SSE files to a separate package
